add green costumer & warnings

+++ src/entity/Tower.cpp
@@ -1,11 +1,11 @@
 #include "Tower.hpp"
 
 bool Tower::isInRange(const Enemy& enemy) const {
-    int distance = std::max(std::abs(x-enemy.x) , std::abs(y - enemy.y));
+    int distance = std::max(std::abs(x-static_cast<int>(enemy.x)) , std::abs(y - static_cast<int>(enemy.y)));
     return distance <= range;
 }
 
-void Tower::attack(std::vector<Enemy> &enemies , float currentTime , std::vector<Projectile> &projectiles)  {
+void Tower::attack(std::vector<Enemy> &enemies , double currentTime , std::vector<Projectile> &projectiles)  {
     for (Enemy &enemy : enemies) {
         if (isInRange(enemy) && currentTime - lastAttackTime >= cadence) {
             lastAttackTime = currentTime;
